+++ packages/app/vitebox/init/main.js
@@ -7,11 +7,23 @@
 import { assert } from './assert.js'
 
 import { initializeApp } from '@firebase/app'
-import { getAuth, connectAuthEmulator } from '@firebase/auth'
+import { getAuth, connectAuthEmulator, debugErrorMap, initializeAuth } from '@firebase/auth'
 import { getFirestore, connectFirestoreEmulator } from '@firebase/firestore'
 import { getFunctions, connectFunctionsEmulator } from '@firebase/functions'
 
 const LOCAL = import.meta.env.MODE === "dev_local";
+
+/*** nah
+// Decide, whether you want human readable error messages - or maybe the normal ones are just fine.
+// This only applies to development. Samples:
+//
+//    true:   "The email address is already in use by another account."
+//    false:  "Firebase: Error (auth/email-already-in-use)."
+//
+// See -> https://github.com/firebase/firebase-js-sdk/issues/5305
+//
+const HUMAN_READABLE_AUTH_ERRORS_PLEASE = true;
+***/
 
 async function initFirebaseLocal(host) {   // (string) => Promise of ()
   assert(LOCAL);
@@ -55,20 +67,24 @@ async function initFirebaseLocal(host) {   // (string) => Promise of ()
   //    on other subpackages). #firebase
   //
   const fns = getFunctions(fah /*, regionOrCustomDomain*/ );
-  const auth = getAuth();
+
+  const auth = /*HUMAN_READABLE_AUTH_ERRORS_PLEASE ? initializeAuth(fah, { errorMap: debugErrorMap })
+    :*/ getAuth();
 
   connectFirestoreEmulator(firestore, host,FIRESTORE_PORT);
   connectFunctionsEmulator(fns, host,FUNCTIONS_PORT);
   connectAuthEmulator(auth, AUTH_URL);
 
+  /*** REMOVE??
   // Signal to Cypress tests that Firebase can be used (emulation setup is done).
   //
   // Note: Was NOT able to do 'getAuth()' on the Cypress side so we pass the auth handle (and whatever else is necessary?)
-  //    from here. Not sure why this is so.
+  //    from here. Not sure why this is so. (likely Cypress code runs before this has?)
   //
   // Importing anything from the app side must be done dynamically.
   //
   window["Let's test!"] = [auth];   // [FirebaseAuth]
+  ***/
 }
 
 /*