+++ packages/backend/test-fns/userInfo.test.js
@@ -6,16 +6,20 @@
 */
 import { test, expect, describe, beforeAll } from '@jest/globals'
 
-import { collection, doc } from 'firebase-jest-testing/firestoreAdmin'
+import { collection, doc, preheat_EXP } from 'firebase-jest-testing/firestoreAdmin'
 
 import './matchers/timesOut'
 import './matchers/toContainObject'
 
+// First call (when server is cold started) takes ~2500 ms (native macOS). To get away from that, DC runs a warm-up
+// lap on this test, reducing times by ~2s.
+
 describe("userInfo shadowing", () => {
 
-  beforeAll( async () => {
-    // Pre-heat also the client (cuts ~320ms off listening times for the stated collection). To show worst case times, skip it.
-    preHeat("projects/1/userInfo");
+  // Warm up the client. Cuts ~300ms from the reported test results (653 -> 367 ms); native macOS
+  //
+  beforeAll( () => {
+    preheat_EXP("projects/1/userInfo");
   })
 
   test('Central user information is distributed to a project where the user is a member', async () => {
@@ -29,9 +33,16 @@ describe("userInfo shadowing", () => {
     await collection("userInfo").doc("abc").set(william);
 
     await expect( docListener("projects/1/userInfo/abc") ).resolves.toContainObject(william);
-  } );    // 340, 388, 406 ms
+  });
+    // DC (mac):
+    //  - no warm-up:   3622, 3795 ms     # run from clean: 'docker compose down', 'docker compose up warm-up'
+    //  - warmed up:     559,  729 ms
+    //
+    // CI (DC):
+    //  - no warm-up:   2184 ms           # warm-up disabled by editing the DC yml
+    //  - warmed up:     550,  678 ms
 
-  test ('Central user information is not distributed to a project where the user is not a member', async () => {
+  test('Central user information is not distributed to a project where the user is not a member', async () => {
 
     // Write in 'userInfo' -> should NOT turn up in project 1.
     //
@@ -57,9 +68,4 @@ function docListener(docPath) {    // (string) => Promise of {...Firestore docum
       /*await*/ unsub();
     });
   });
-}
-
-function preHeat(docPath) {    // (string) => ()
-  const unsub = doc(`${docPath}/...`).onSnapshot( ss => {} );
-  unsub();
 }