+++ sh/run-no.sh
@@ -8,8 +8,14 @@ set -eu -o pipefail
 #     $ sh/run-no.sh
 #   <<
 #
+PATHS=". tools dc-tools/firebase-prime/package packages/backend packages/backend/functions packages/app packages/app-deploy-ops"
 
-PATHS=". tools firebase-ci-builder.user/firebase-prime packages/backend packages/backend/functions packages/app packages/app-deploy-ops"
+# Compensate for the macOS/Linux (Docker) modifications of 'esbuild'. Otherwise:
+#   <<
+#     npm ERR! Cannot set property 'peer' of null
+#   <<
+#
+npm --prefix packages/app prune
 
 for _PATH in $PATHS
 do